requirements.txt
bcrypt==4.2.0
boto3==1.34.151
botocore==1.34.151
cffi==1.16.0
cryptography==43.0.0
jmespath==1.0.1
paramiko==3.4.0
pycparser==2.22
PyNaCl==1.5.0
pysftp==0.2.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.2
six==1.16.0
urllib3==2.2.2
